pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Konstantin Krivosheev",email = "Krivosheev2008@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "apache-airflow (==3.1.7)",
    "pandas (>=3.0.0,<4.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "apache-airflow-providers-mongo (>=5.3.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
